Refactor Go DTO parser
