+++ crates/rulemorph_mcp/src/dto_parse.rs
@@ -7,6 +7,10 @@ use crate::dto_normalize::{
 };
 use crate::dto_schema::{DtoField, DtoFieldType, DtoSchema, DtoType, PrimitiveKind};
 
+mod go;
+
+use self::go::parse_go_types;
+
 pub(crate) fn parse_dto_schema(
     text: &str,
     language: DtoSourceLanguage,
@@ -461,208 +465,6 @@ fn parse_python_types(text: &str) -> Result<(HashMap<String, DtoType>, Vec<Strin
     }
 
     Ok((types, order))
-}
-
-fn parse_go_types(text: &str) -> Result<(HashMap<String, DtoType>, Vec<String>), String> {
-    let mut types: HashMap<String, DtoType> = HashMap::new();
-    let mut order = Vec::new();
-    let mut index = 0usize;
-    let bytes = text.as_bytes();
-
-    while index < bytes.len() {
-        let slice = &text[index..];
-        let Some(pos) = slice.find("type ") else {
-            break;
-        };
-        index += pos + 5;
-        let rest = &text[index..];
-        let name = rest.trim_start().split_whitespace().next().unwrap_or("");
-        if name.is_empty() {
-            index = index.saturating_add(1);
-            continue;
-        }
-        let name_start = rest.find(name).unwrap_or(0);
-        index += name_start + name.len();
-        let after_name = &text[index..];
-        let Some(struct_pos) = after_name.find("struct") else {
-            index = index.saturating_add(1);
-            continue;
-        };
-        index += struct_pos + "struct".len();
-        let after_struct = &text[index..];
-        let Some(brace_pos) = after_struct.find('{') else {
-            index = index.saturating_add(1);
-            continue;
-        };
-        index += brace_pos + 1;
-
-        let mut brace_depth = 1usize;
-        let mut body_end = index;
-        while body_end < bytes.len() {
-            match bytes[body_end] as char {
-                '{' => brace_depth += 1,
-                '}' => {
-                    brace_depth -= 1;
-                    if brace_depth == 0 {
-                        break;
-                    }
-                }
-                _ => {}
-            }
-            body_end += 1;
-        }
-        if brace_depth != 0 {
-            break;
-        }
-
-        let body = &text[index..body_end];
-        let dto_type = types
-            .entry(name.to_string())
-            .or_insert_with(|| DtoType { fields: Vec::new() });
-        parse_go_struct_fields(body, dto_type);
-        order.push(name.to_string());
-        index = body_end + 1;
-    }
-
-    Ok((types, order))
-}
-
-fn parse_go_struct_fields(body: &str, dto_type: &mut DtoType) {
-    let mut chars = body.chars().peekable();
-    while let Some(ch) = chars.peek().copied() {
-        if ch.is_whitespace() {
-            chars.next();
-            continue;
-        }
-        if ch == '/' {
-            chars.next();
-            if matches!(chars.peek(), Some('/')) {
-                while let Some(next) = chars.next() {
-                    if next == '\n' {
-                        break;
-                    }
-                }
-                continue;
-            }
-            if matches!(chars.peek(), Some('*')) {
-                chars.next();
-                while let Some(next) = chars.next() {
-                    if next == '*' && matches!(chars.peek(), Some('/')) {
-                        chars.next();
-                        break;
-                    }
-                }
-                continue;
-            }
-            continue;
-        }
-
-        let field_name = read_go_token(&mut chars);
-        if field_name.is_empty() {
-            chars.next();
-            continue;
-        }
-        skip_go_whitespace(&mut chars);
-        let field_type = read_go_token(&mut chars);
-        if field_type.is_empty() {
-            continue;
-        }
-
-        skip_go_whitespace(&mut chars);
-        let tag = if matches!(chars.peek(), Some('`')) {
-            chars.next();
-            let mut tag_value = String::new();
-            while let Some(next) = chars.next() {
-                if next == '`' {
-                    break;
-                }
-                tag_value.push(next);
-            }
-            Some(tag_value)
-        } else {
-            None
-        };
-
-        let (json_key, tag_optional, skip_field) = parse_go_json_tag(tag.as_deref());
-        if skip_field {
-            continue;
-        }
-
-        let mut optional = tag_optional;
-        let mut type_token = field_type.trim().to_string();
-        if let Some(stripped) = type_token.strip_prefix('*') {
-            optional = true;
-            type_token = stripped.to_string();
-        }
-
-        let field_type = if type_token.contains('[') || type_token.contains("map[") {
-            DtoFieldType::Unknown
-        } else {
-            match type_token.as_str() {
-                "string" => DtoFieldType::Primitive(PrimitiveKind::String),
-                "bool" => DtoFieldType::Primitive(PrimitiveKind::Bool),
-                "int" | "int8" | "int16" | "int32" | "int64" | "uint" | "uint8" | "uint16"
-                | "uint32" | "uint64" | "uintptr" => DtoFieldType::Primitive(PrimitiveKind::Int),
-                "float32" | "float64" => DtoFieldType::Primitive(PrimitiveKind::Float),
-                "" => DtoFieldType::Unknown,
-                other => DtoFieldType::Object(other.to_string()),
-            }
-        };
-
-        let json_key = json_key.unwrap_or_else(|| field_name.clone());
-        dto_type.fields.push(DtoField {
-            json_key,
-            field_type,
-            optional,
-        });
-    }
-}
-
-fn read_go_token(chars: &mut std::iter::Peekable<std::str::Chars<'_>>) -> String {
-    let mut token = String::new();
-    while let Some(&ch) = chars.peek() {
-        if ch.is_whitespace() || ch == '`' || ch == '{' || ch == '}' {
-            break;
-        }
-        token.push(ch);
-        chars.next();
-    }
-    token
-}
-
-fn skip_go_whitespace(chars: &mut std::iter::Peekable<std::str::Chars<'_>>) {
-    while let Some(&ch) = chars.peek() {
-        if !ch.is_whitespace() {
-            break;
-        }
-        chars.next();
-    }
-}
-
-fn parse_go_json_tag(tag: Option<&str>) -> (Option<String>, bool, bool) {
-    let Some(tag) = tag else {
-        return (None, false, false);
-    };
-    let Some(start) = tag.find("json:\"") else {
-        return (None, false, false);
-    };
-    let after = &tag[start + 6..];
-    let Some(end) = after.find('"') else {
-        return (None, false, false);
-    };
-    let content = &after[..end];
-    if content == "-" {
-        return (None, false, true);
-    }
-    let mut parts = content.split(',');
-    let name_part = parts.next().unwrap_or("");
-    let omitempty = parts.any(|part| part.trim() == "omitempty");
-    let name = if name_part.is_empty() {
-        None
-    } else {
-        Some(name_part.to_string())
-    };
-    (name, omitempty, false)
 }
 
 fn parse_java_types(text: &str) -> Result<(HashMap<String, DtoType>, Vec<String>), String> {
